+++ test/migrations/20241125000010_create_chat_events.exs
@@ -1,0 +1,17 @@
+defmodule SpectorTest.Repo.Migrations.CreateAIChatEvents do
+  use Ecto.Migration
+
+  def up do
+    Spector.Migration.up(
+      table: "ai_chat_events",
+      links: [{"ai_chat_ancestors", :ancestor_id}]
+    )
+  end
+
+  def down do
+    Spector.Migration.down(
+      table: "ai_chat_events",
+      links: [{"ai_chat_ancestors", :ancestor_id}]
+    )
+  end
+end