+++ test/migrations/20241125000011_create_basic_chat_events.exs
@@ -1,0 +1,17 @@
+defmodule SpectorTest.Repo.Migrations.CreateBasicChatEvents do
+  use Ecto.Migration
+
+  def up do
+    Spector.Migration.up(
+      table: "basic_chat_events",
+      links: [{"basic_chat_edits", :previous_id}]
+    )
+  end
+
+  def down do
+    Spector.Migration.down(
+      table: "basic_chat_events",
+      links: [{"basic_chat_edits", :previous_id}]
+    )
+  end
+end